Expliclity expunge objects from session
Since we are using a thread local session, if we want to pass on
objects, we should make sure to expunge (i.e. disconect) them from the
session before passing them on. This code makes sure that this is done
expliclity for the type of methods on the job repo where I though that
it would make sense to do so.