Configure alembic programatically
Previously a mix of values from the alemtic.ini file and what was
specified in the app config were used. This was confusing. I therefore
made sure that for the app all config for this is in app.config. The
alembic.ini file is kept around for convenience since this is necessary
for autogenerating alembic update scripts.