Start Nextflow with shell=True
This appears to fix the weird permission problems we have been seeing on
the staging system. My guess as to why this works is that some part of
Nextflow relies on being executes in a bash shell, which a normal POpen
python process is not. This fix here, however does not execute a bash
shell, but /bin/sh which defaults to /bin/dash on the sytems I looked
at, so technically I guess it shouldn't work. But I tried it on staging
and seems to work... All in all, if this fixes the problem I'm happy,
but I would have liked to understand better why it actually works...