Add support for nf-core input samplesheets
Most nf-core pipelines define input data in a pipeline-specific
samplesheet. This commit makes it possible to provide such a samplesheet
to the pipeline.

The content of the samplesheet can be either provided directly to the
endpoint or can be described in the config file (in this case it will
usually make use of {runfolder_name} and {runfolder_path}.